Revert "Return JSON-LD in HTTP 4xx responses (#555)"
This reverts commit c9aac7550893a65a1fd5b35c7a796fab644bcb7c.
