+++ CHANGES.md
@@ -2,7 +2,6 @@
 
 ## 5.0.6
 
-* IIIF information endpoints always return JSON in HTTP 4xx responses.
 * Fixed a bug whereby the values of the `operations` and `page_count` keys
   in the delegate context were not set.
 * TurboJpegProcessor is able to generate non-JPEG derivative images, which
