+++ src/main/java/edu/illinois/library/cantaloupe/resource/iiif/v1/InformationResource.java
@@ -1,9 +1,7 @@
 package edu.illinois.library.cantaloupe.resource.iiif.v1;
 
 import java.util.HashSet;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
 import edu.illinois.library.cantaloupe.http.Method;
@@ -54,7 +52,16 @@ public class InformationResource extends IIIF1Resource {
         class CustomCallback implements InformationRequestHandler.Callback {
             @Override
             public boolean authorize() throws Exception {
-                return InformationResource.this.preAuthorize();
+                try {
+                    // The logic here is somewhat convoluted. See the method
+                    // documentation for more information.
+                    return InformationResource.this.preAuthorize();
+                } catch (ResourceException e) {
+                    if (e.getStatus().getCode() > 400) {
+                        throw e;
+                    }
+                }
+                return false;
             }
             @Override
             public void knowAvailableOutputFormats(Set<Format> formats) {
@@ -70,25 +77,18 @@ public class InformationResource extends IIIF1Resource {
                 .withRequestContext(getRequestContext())
                 .withCallback(new CustomCallback())
                 .build()) {
-            try {
-                Info info = handler.handle();
-                ImageInfo iiifInfo = new ImageInfoFactory().newImageInfo(
-                        getImageURI(),
-                        availableOutputFormats,
-                        info,
-                        getPageIndex(),
-                        getMetaIdentifier().getScaleConstraint());
-                addHeaders(iiifInfo);
-                new JacksonRepresentation(iiifInfo)
-                        .write(getResponse().getOutputStream());
-            } catch (ResourceException e) {
-                if (e.getStatus().getCode() < 500) {
-                    newHTTP4xxRepresentation(e.getStatus(), e.getMessage())
-                            .write(getResponse().getOutputStream());
-                } else {
-                    throw e;
-                }
-            }
+            Info info = handler.handle();
+
+            ImageInfo iiifInfo = new ImageInfoFactory().newImageInfo(
+                    getImageURI(),
+                    availableOutputFormats,
+                    info,
+                    getPageIndex(),
+                    getMetaIdentifier().getScaleConstraint());
+
+            addHeaders(iiifInfo);
+            new JacksonRepresentation(iiifInfo)
+                    .write(getResponse().getOutputStream());
         }
     }
 
@@ -120,16 +120,6 @@ public class InformationResource extends IIIF1Resource {
             mediaType = "application/json";
         }
         return mediaType + ";charset=UTF-8";
-    }
-
-    private JacksonRepresentation newHTTP4xxRepresentation(Status status,
-                                                           String message) {
-        final Map<String, Object> map = new LinkedHashMap<>(); // preserves key order
-        map.put("@context", "http://library.stanford.edu/iiif/image-api/1.1/context.json");
-        map.put("@id", getImageURI());
-        map.put("status", status.getCode());
-        map.put("message", message);
-        return new JacksonRepresentation(map);
     }
 
 }
