+++ src/test/java/edu/illinois/library/cantaloupe/resource/iiif/v2/InformationResourceTest.java
@@ -51,22 +51,7 @@ public class InformationResourceTest extends ResourceTest {
     @Test
     void testGETAuthorizationWhenUnauthorized() {
         URI uri = getHTTPURI("/unauthorized.jpg/info.json");
-        // This may vary depending on the return value of a delegate method,
-        // but the test delegate script returns 401.
-        assertStatus(401, uri);
-        assertRepresentationEquals("{\"@context\":\"http://iiif.io/api/image/2/context.json\","+
-                "\"@id\":\"" + uri.toString().replace("/info.json", "") + "\"," +
-                "\"protocol\":\"http://iiif.io/api/image\"," +
-                "\"status\":401," +
-                "\"message\":\"Unauthorized\"," +
-                "\"attribution\":\"Copyright My Great Organization. All rights reserved.\"," +
-                "\"license\":\"http://example.org/license.html\"," +
-                "\"service\":{" +
-                    "\"@context\":\"http://iiif.io/api/annex/services/physdim/1/context.json\"," +
-                    "\"profile\":\"http://iiif.io/api/annex/services/physdim\"," +
-                    "\"physicalScale\":0.0025," +
-                    "\"physicalUnits\":\"in\"}" +
-                "}", uri);
+        tester.testAuthorizationWhenUnauthorized(uri);
     }
 
     @Test
